requirements.txt
a2a-sdk
dotenv
google-adk
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-secret-manager
langchain
langgraph
langserve
